Fix all errors in new Raknet crate
