+++ core/src/network/interaction.rs
