+++ core/src/network/manager.rs
