--- core/src/network/session/mod.rs
@@ -1,9 +1,0 @@
-use ::util::glob_export;
-
-glob_export!(level);
-glob_export!(session);
-glob_export!(manager);
-glob_export!(login);
-glob_export!(interaction);
-glob_export!(handlers);
-
