+++ raknet/Cargo.toml
@@ -3,6 +3,14 @@ name = "pyro-raknet"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
+util = { package = "pyro-util", path = "../util" }
+proto = { package = "pyro-proto", path = "../protocol" }
+
+tracing = "0.1.40"
+tokio = { version = "1.35.1", features = ["sync"] }
+tokio-util = "0.7.10"
+async-recursion = "1.0.5"
+anyhow = "1.0.79"
+dashmap = "5.5.3"
+parking_lot = "0.12.1"
