--- raknet/src/mod.rs
@@ -1,18 +1,0 @@
-//! Rust implementation of the RakNet protocol.
-
-use util::glob_export;
-
-glob_export!(ack);
-glob_export!(broadcast);
-glob_export!(compound);
-glob_export!(frame);
-glob_export!(login);
-glob_export!(order);
-glob_export!(raw);
-glob_export!(receive);
-glob_export!(recovery);
-glob_export!(reliability);
-glob_export!(send_queue);
-glob_export!(send);
-glob_export!(session);
-glob_export!(tick);
