+++ raknet/src/recovery.rs
@@ -1,7 +1,7 @@
 use dashmap::DashMap;
 use proto::raknet::AckRecord;
 
-use crate::raknet::FrameBatch;
+use crate::FrameBatch;
 
 /// Holds previously sent raknet to be able to recover them when packet loss occurs.
 ///
