+++ raknet/src/send.rs
@@ -2,8 +2,6 @@ use std::io::Write;
 use std::sync::atomic::Ordering;
 
 use async_recursion::async_recursion;
-use flate2::Compression;
-use flate2::write::DeflateEncoder;
 use proto::bedrock::{CompressionAlgorithm, CONNECTED_PACKET_ID, ConnectedPacket, Packet};
 use proto::raknet::{Ack, AckRecord};
 
@@ -11,11 +9,7 @@ use util::{BinaryWrite, MutableBuffer};
 
 use util::Serialize;
 
-use crate::network::{RaknetUser, BedrockUserLayer};
-use crate::raknet::{Frame, FrameBatch};
-use crate::raknet::Reliability;
-use crate::raknet::SendPriority;
-use crate::config::SERVER_CONFIG;
+use crate::{SendPriority, RaknetUser, Reliability, Frame, FrameBatch};
 
 /// Specifies the reliability and priority of a packet.
 pub struct PacketConfig {
@@ -287,65 +281,5 @@ impl RaknetUser {
         }
 
         compound
-    }
-}
-
-impl BedrockUserLayer {
-    /// Sends a game packet with default settings
-    /// (reliable ordered and medium priority)
-    pub fn send<T: ConnectedPacket + Serialize>(&self, packet: T) -> anyhow::Result<()> {
-        let packet = Packet::new(packet);
-        let serialized = packet.serialize()?;
-
-        self.send_serialized(serialized, DEFAULT_SEND_CONFIG)
-    }
-
-    /// Sends a game packet with custom reliability and priority
-    pub fn send_serialized<B>(&self, packet: B, config: PacketConfig) -> anyhow::Result<()>
-        where
-            B: AsRef<[u8]>
-    {
-        let mut out;
-        if self.compressed.get() {
-            let (algorithm, threshold) = {
-                let config = SERVER_CONFIG.read();
-                (config.compression_algorithm, config.compression_threshold)
-            };
-
-            if packet.as_ref().len() > threshold as usize {
-                // Compress packet
-                match algorithm {
-                    CompressionAlgorithm::Snappy => {
-                        unimplemented!("Snappy compression");
-                    }
-                    CompressionAlgorithm::Deflate => {
-                        let mut writer = DeflateEncoder::new(
-                            vec![CONNECTED_PACKET_ID],
-                            Compression::best(),
-                        );
-
-                        writer.write_all(packet.as_ref())?;
-                        out = MutableBuffer::from(writer.finish()?)
-                    }
-                }
-            } else {
-                // Also reserve capacity for checksum even if encryption is disabled,
-                // preventing allocations.
-                out = MutableBuffer::with_capacity(1 + packet.as_ref().len() + 8);
-                out.write_u8(CONNECTED_PACKET_ID)?;
-                out.write_all(packet.as_ref())?;
-            }
-        } else {
-            // Also reserve capacity for checksum even if encryption is disabled,
-            // preventing allocations.
-            out = MutableBuffer::with_capacity(1 + packet.as_ref().len() + 8);
-            out.write_u8(CONNECTED_PACKET_ID)?;
-            out.write_all(packet.as_ref())?;
-        };
-
-        self.encryptor.encrypt(&mut out)?;
-
-        self.raknet.send_raw_buffer_with_config(out, config);
-        Ok(())
     }
 }
